Optionally let client_request test for redirect
A common pattern we employ is `POST`-redirect-`GET`. To write tests for
this we often check that the URL of the redirect is what we’re
expecting.

With the way `client_request` is currently set up, there’s no way to do
this because the response isn’t exposed to the test.

So this commit adds an extra parameter which will let us test for
redirects.