Test existing functionality
When picking an API key you get a choice of different key types. Which
types are available depends on whether or not your service is in trial
mode.

This commit adds missing tests for the existing functionality which
determined the types of key that should be available to the user.