Fix invisible focused item text
GOV.UK Template hanged the colour of text in focused links in
https://github.com/alphagov/govuk_template/commit/79466a489c38b0b82f4cb95daa2baba41b375a81

It was done with greater specificity than before. This means that the
way we were previously overriding the focus colour (for links with a
dark background) no longer works.

This commit makes our override more specific, so that it works again.