Make clickable area of link bigger
Fitt’s law[1] states that bigger click areas are quicker and easier for
people to click. Therefore we should make click areas as big as
possible, without being ambiguous about what the outcome of clicking
will be or increasing the potential for accidental clicks.

The click areas of the row numbers in the table were very small – this
commits makes them as big as the containing table cells.

Uses this technique to achieve the bigger click areas without disrupting
the layout:
http://authenticff.com/journal/css-pro-tip-expanding-clickable-area

1. https://en.wikipedia.org/wiki/Fitts%27s_law