Put ‘add new’ button at top of page
One of the most frequent tasks done on this page is adding a new
branding.

Current to do this you have to:
- scroll to the bottom
- scan for the ‘Create a new email branding’ option that visually looks
  just like all the other brandings
- submit a form

This commit makes change it to one clearly differentiated button at the
top of the page. This is consistent for how we let users add templates
and team members.