Add tests for changes to brand request page
Tests the new code that gets the brand type from
the email_branding model. Includes checks for a
service without the email_branding field set.

It also amends the test for a POST from that page,
removing mocking of the email_branding client.
This test runs against the default service which
has its email_branding field set to None so no
call is made to the client. It's testing the
brand_type values selected so doesn't need the
service to have an email_branding already set.