Commented out the link to documentation
Commented out the link to documentation because sending a document is only described in the Python documentation at the moment.

Discussed with Chris H:
Added a line asking users to contact the team. Support can explain the situation if we get any requests. We'll remove this and reinstate the original line once this feature is available across more API clients and the documentation is updated.